+++ src/app/board/page.jsx
@@ -1,0 +1,3 @@
+export default function BoardListPage() {
+  return <div></div>;
+}
